facade/requirements.txt
django
requests
celery
hazelcast-python-client
python-consul